# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/me": {
      "get": {
        "operationId": "UsersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Current customer profile with saved addresses",
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update profile (name/email)",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/fcm-token": {
      "put": {
        "operationId": "UsersController_saveFcmToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FcmTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register this device for push notifications (customer or rider token)",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/addresses": {
      "post": {
        "operationId": "UsersController_createAddress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a delivery address (must carry a zoneId from GET /v1/zones)",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/addresses/{id}": {
      "patch": {
        "operationId": "UsersController_updateAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit a saved address",
        "tags": [
          "me"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a saved address",
        "tags": [
          "me"
        ]
      }
    },
...
  },
  "info": {
    "title": "MediLocal API",
    "description": "Hyperlocal medicine delivery platform API. The OpenAPI spec at /docs-json also feeds the Dart client generator for the Flutter apps.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Ravi Kumar"
          },
          "email": {
            "type": "string",
            "example": "ravi@example.com"
          }
        }
      },
      "FcmTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "FCM device registration token"
          }
        },
        "required": [
          "token"
        ]
      },
      "CreateAddressDto": {
        "type": "object",
        "properties": {
          "zoneId": {
            "type": "string",
            "description": "Delivery zone the address falls in (from GET /v1/zones)"
          },
          "label": {
            "type": "string",
            "example": "Home",
            "default": "Home"
          },
          "line1": {
            "type": "string",
            "example": "H.No 12, Gandhi Nagar"
          },
          "line2": {
            "type": "string"
          },
          "landmark": {
            "type": "string",
            "example": "Opposite SBI ATM"
          },
          "pincode": {
            "type": "string",
            "example": "800001"
          },
          "lat": {
            "type": "number",
            "example": 25.5941,
            "description": "Map-pin latitude"
          },
          "lng": {
            "type": "number",
            "example": 85.1376,
            "description": "Map-pin longitude"
          }
        },
        "required": [
          "zoneId",
          "line1",
          "lat",
          "lng"
        ]
      },
      "UpdateAddressDto": {
        "type": "object",
        "properties": {
          "zoneId": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "line1": {
            "type": "string"
          },
          "line2": {
            "type": "string"
          },
          "landmark": {
            "type": "string"
          },
          "pincode": {
            "type": "string"
          },
          "lat": {
            "type": "number"
          },
          "lng": {
            "type": "number"
          }
        }
      },
...
    }
  }
}